test(agent): split external workspace contract
