+++ framework/core/tests/python/test_agent_console_ambient_adoption.py
@@ -1,11 +1,14 @@
 # SPDX-License-Identifier: Apache-2.0
 
+import json
 from types import SimpleNamespace
 
 from click.testing import CliRunner
 
 from kungfu.agent import run_agent, session_contract, session_surface
 from kungfu.cli.commands import kfc
+from kungfu.workspace import resolve_workspace_target
+from agent_bootstrap_fixtures import verified_bootstrap_receipt
 
 
 ROOT_HASH = "sha256:" + "a" * 64
@@ -183,3 +186,117 @@ def test_public_bind_work_cli_adopts_current_codex_process(monkeypatch, tmp_path
     assert calls[0][1]["adopt"] is True
     assert calls[1][2]["envelope_override"] == envelope
     assert calls[1][2]["console_workspace_root"] == str(tmp_path)
+
+
+def test_public_bind_work_cli_selects_explicit_external_project(monkeypatch, tmp_path):
+    requests = []
+    observed_runtime_dirs = []
+    console_project = tmp_path / "console-project"
+    console_runtime = console_project / ".kungfu" / "runtime"
+    console_runtime.mkdir(parents=True)
+    console_target = resolve_workspace_target(
+        "read-only", str(console_project), cwd=str(console_project)
+    )
+    work_project = tmp_path / "work-project"
+    work_runtime = work_project / ".kungfu" / "runtime"
+    work_runtime.mkdir(parents=True)
+    work_target = resolve_workspace_target(
+        "read-only", str(work_project), cwd=str(work_project)
+    )
+    body = {
+        "schema": "kungfu.agent-console-envelope/v1",
+        "workspaceId": console_target.identity.workspace_id,
+        "consoleId": f"assistant:{console_target.identity.workspace_id}:native:one",
+        "attemptId": "native:one",
+        "runtimeProfileId": "kungfu.agent-runtime.codex.test",
+        "provider": "codex",
+        "activeProfiles": [],
+        "workRef": None,
+        "entrypoints": {
+            "context": ["/exact/kungfu", "agent", "context", "--json"],
+            "capabilities": [
+                "/exact/kungfu",
+                "agent",
+                "capabilities",
+                "--json",
+            ],
+            "profiles": ["/exact/kungfu", "profile", "manager", "--json"],
+            "bindWork": ["/exact/kungfu", "agent", "console", "bind-work"],
+        },
+        "knownLimits": [],
+    }
+    envelope = {**body, "envelopeRoot": run_agent.canonical_root(body)}
+    monkeypatch.setenv("KUNGFU_AGENT_CONSOLE_ENVELOPE", json.dumps(envelope))
+    monkeypatch.setenv(
+        "KUNGFU_AGENT_BOOTSTRAP_RECEIPT",
+        json.dumps(verified_bootstrap_receipt()),
+    )
+    monkeypatch.setenv("KUNGFU_AGENT_ATTEMPT_ID", "native:one")
+    monkeypatch.setenv("KUNGFU_WORKSPACE_ROOT", str(console_project))
+    monkeypatch.setenv("KUNGFU_AGENT_RUNTIME_DIR", str(console_runtime))
+
+    def status(runtime_dir, *_args):
+        observed_runtime_dirs.append(runtime_dir)
+        return {
+            "assignment": {"assignment_id": "assignment:external"},
+            "query_proof_root": ROOT_HASH,
+        }
+
+    monkeypatch.setattr("kungfu.cli.commands.assignment._status", status)
+    monkeypatch.setattr(
+        "kungfu.cli.commands.assignment.profile_source", lambda: tmp_path
+    )
+
+    def validate_source(_source, runtime_dir):
+        observed_runtime_dirs.append(runtime_dir)
+        return {
+            "inspection": {
+                "profile": {"id": "kungfu.work-control"},
+                "profile_suite_root": ROOT_HASH,
+            }
+        }
+
+    monkeypatch.setattr("kungfu.profile_sdk.validate_source", validate_source)
+
+    def invoke(request, **_kwargs):
+        requests.append(request)
+        if request["operation"] == "plan-native-bind-work":
+            return {
+                "operation": "native-bind-work",
+                "root": ROOT_HASH,
+                **request["input"]["session"],
+                "workRef": request["input"]["workRef"],
+            }
+        return {"status": "bound", "receiptRoot": ROOT_HASH}
+
+    monkeypatch.setattr(run_agent.session_surface, "invoke", invoke)
+    result = CliRunner().invoke(
+        kfc,
+        [
+            "--home",
+            str(tmp_path / "kungfu-home"),
+            "agent",
+            "console",
+            "bind-work",
+            "--initiative-id",
+            "initiative:external",
+            "--assignment-id",
+            "assignment:external",
+            "--workspace",
+            str(work_project),
+            "--json",
+        ],
+    )
+
+    assert result.exit_code == 0, result.output
+    payload = json.loads(result.output)
+    assert payload["status"] == "bound"
+    assert payload["workRef"]["workspaceId"] == work_target.identity.workspace_id
+    assert payload["session"]["workConsoleId"] == envelope["consoleId"]
+    assert observed_runtime_dirs == [str(work_runtime), str(work_runtime)]
+    assert requests[0]["input"]["bindingScope"] == "explicit-external-project"
+    assert requests[0]["input"]["sourceWorkspaceId"] == envelope["workspaceId"]
+    assert [request["operation"] for request in requests] == [
+        "plan-native-bind-work",
+        "bind-native-work",
+    ]
